+++ lib/hashie/extensions/deep_find.rb
@@ -27,32 +27,15 @@ module Hashie
       private
 
       def _deep_find(key, object = self)
-        if object.respond_to?(:key?)
-          return object[key] if object.key?(key)
-
-          reduce_to_match(key, object.values)
-        elsif object.is_a?(Enumerable)
-          reduce_to_match(key, object)
-        end
+        _deep_find_all(key, object).first
       end
 
       def _deep_find_all(key, object = self, matches = [])
-        if object.respond_to?(:key?)
-          matches << object[key] if object.key?(key)
-          object.values.each { |v| _deep_find_all(key, v, matches) }
-        elsif object.is_a?(Enumerable)
-          object.each { |v| _deep_find_all(key, v, matches) }
+        deep_locate_result = Hashie::Extensions::DeepLocate.deep_locate(key, object).tap do |result|
+          result.map! { |element| element[key] }
         end
 
-        matches
-      end
-
-      def reduce_to_match(key, enumerable)
-        enumerable.reduce(nil) do |found, value|
-          return found if found
-
-          _deep_find(key, value)
-        end
+        matches.concat(deep_locate_result)
       end
     end
   end